:recycle: refactor: Ajustes repetição de código
