+++ src/components/RecargaAutomatica.jsx
@@ -5,11 +5,10 @@ import "../pages/Style/styles.css";
 import FuncTrocaComp from "../util/FuncTrocaComp";
 import VoltarComponente from "../util/VoltarComponente";
 import { CiEdit } from "react-icons/ci";
-import { Group, Text, Button, Image, Box, Divider, Modal } from "@mantine/core";
-import { BsCreditCard2Back } from "react-icons/bs";
-import createAPI from "../services/createAPI";
-import { IconCheck, IconX } from "@tabler/icons-react";
+import { Button, Divider, Modal } from "@mantine/core";
 
+import { IconCheck, IconX } from "@tabler/icons-react";
+import ListarCartao from "../util/ListarCartao";
 const RecargaAutomatica = () => {
     const [recargaAtivada, setRecargaAtivada] = useState(true);
     const [saldoLimite, setSaldoLimite] = useState();
@@ -22,34 +21,7 @@ const RecargaAutomatica = () => {
     const [CreditCardSelected, setCreditCardSelected] = useState(null);
     const [opened, { open, close }] = useDisclosure(false);
     const [estadoDiv, setEstadoDiv] = useState(false);
-const getCreditCardFUNC = async () => {
-    const requisicao = createAPI();
 
-    await requisicao
-        .get("/cartao/")
-        .then((resposta) => {
-            if (resposta.data.msg.resultado) {
-                const newData = resposta.data.data.map((item) => ({
-                    cartao: item.id_cartao,
-                    bandeira: item.bandeira,
-                    numero: `#### #### #### ${item.cartao}`,
-                    debito: item.debito,
-                    credito: item.credito,
-                }));
-                setCreditCard(newData);
-            } else {
-                setCreditCard([]);
-                FuncTrocaComp('CartaoCredito');
-            }
-        })
-        .catch((err) => {
-            setCreditCard([]);
-        });
-};
-
-useEffect(() => {
-    getCreditCardFUNC();
-}, []);
 
     const handleToggle = () => {
         if (!recargaAtivada) {
@@ -96,7 +68,9 @@ useEffect(() => {
         setValorRecarga(valorRecargaInputValue);
         setIsValorRecargaButtonVisible(true);
     }
-
+    const handleNoCards = () => {
+        FuncTrocaComp('CartaoCredito');
+      };
 
     return (
         <div>
@@ -197,63 +171,7 @@ useEffect(() => {
                                 <Button className='botao-add mt-3 py-3 text-white' style={{ backgroundColor: '#1099A8' }} onClick={() => { FuncTrocaComp('HistoricoRecargaAuto'); }}>
                                     Ver Histórico
                                 </Button>
-                                <Group className='mt-4 mb-3' position="apart">
-                                    <Text weight={500}>Selecione seu cartão:</Text>
-                                </Group> {creditCard.map((item, index) => (
-                                    (
-                                        <div key={index}>
-                                            <Box className="border border-black rounded mb-2 align-items-center"
-                                                style={{ maxWidth: '400px', minHeight: '50px', backgroundImage: CreditCardSelected === index ? 'linear-gradient(45deg, #0CA678,  #1098AD)' : 'none' }}
-                                                onClick={() => setCreditCardSelected(index)}
-                                            >
-                                                <div className="d-flex align-items-center">
-                                                    <div className="align-items-center">
-                                                        {item.bandeira === 'visa' ? (
-                                                            CreditCardSelected !== index ? (
-                                                                <Image
-                                                                    src='../../assets/img/cartaoCredito/visa-unselected.png'
-                                                                    alt="image"
-                                                                    style={{ width: 45, height: 45, display: 'flex', alignItems: 'center', marginLeft: '5px' }} />
-                                                            ) : (
-                                                                <Image
-                                                                    src='../../assets/img/cartaoCredito/visa.png'
-                                                                    alt="image"
-                                                                    style={{ width: 45, height: 45, display: 'flex', alignItems: 'center', marginLeft: '5px' }} />
-                                                            )
-                                                        ) : item.bandeira === 'mastercard' ? (
-                                                            <Image
-                                                                src='../../assets/img/cartaoCredito/mastercard.png'
-                                                                alt="image"
-                                                                style={{ width: 50, height: 50, display: 'flex', alignItems: 'center' }} />
-                                                        ) : item.bandeira === 'elocard' ? (
-                                                            CreditCardSelected === index ? (
-                                                                <Image
-                                                                    src='../../assets/img/cartaoCredito/elocard.png'
-                                                                    alt="image"
-                                                                    style={{ width: 50, height: 50, display: 'flex', alignItems: 'center', marginLeft: '2px' }} />
-                                                            ) : (
-                                                                <Image
-                                                                    src='../../assets/img/cartaoCredito/elocard-unselected.png'
-                                                                    alt="image"
-                                                                    style={{ width: 50, height: 50, display: 'flex', alignItems: 'center', marginLeft: '2px' }} />
-                                                            )
-                                                        )
-                                                            :
-                                                            <BsCreditCard2Back className="m-2"
-                                                                size={30}
-                                                                color={CreditCardSelected === index ? 'white' : 'black'} />}
-                                                    </div>
-                                                    <div className={CreditCardSelected === index ? "text-start text-white mx-2" : "text-start mx-2"}>
-                                                        <Text weight={400} style={{ marginTop: '-3px' }}>{item.numero}</Text>
-                                                    </div>
-                                                </div>
-                                            </Box>
-                                        </div>
-                                    )
-                                ))}
-                                <Button className='botao-add mt-1 py-3' onClick={() => { FuncTrocaComp('CartaoCredito'); }}>
-                                    Adicionar novo cartão
-                                </Button>
+                                <ListarCartao  metodo={'cartaoCred'} onNoCards={handleNoCards} />
                             </div>
                         </div>
                     </div>
